refactor folders and update javadoc
