+++ src/main/scala/main.scala
@@ -1,7 +1,8 @@
 package fr.maxime.binandco
 
+import tools.Tools
 import tools.Tools.organizeTextIn64BytesBlocksWithLengthEndLine
-import tools.{Sha256, Tools}
+import tools.sha.Sha256
 
 import java.nio.charset.Charset
 import scala.util.control.Breaks.{break, breakable}
